// apps/mobile/http/generated/models/PostOrganizationsSlugSellers.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type PostOrganizationsSlugSellersPathParams = {
  /**
   * @type string
   */
  slug: string;
};

/**
 * @description Default Response
 */
export type PostOrganizationsSlugSellers201 = {
  /**
   * @type string, uuid
   */
  sellerId: string;
};

export const postOrganizationsSlugSellersMutationRequestDocumentTypeEnum = {
  CPF: "CPF",
  CNPJ: "CNPJ",
} as const;

export type PostOrganizationsSlugSellersMutationRequestDocumentTypeEnumKey =
  (typeof postOrganizationsSlugSellersMutationRequestDocumentTypeEnum)[keyof typeof postOrganizationsSlugSellersMutationRequestDocumentTypeEnum];

export const postOrganizationsSlugSellersMutationRequestStatusEnum = {
  ACTIVE: "ACTIVE",
  INACTIVE: "INACTIVE",
} as const;

export type PostOrganizationsSlugSellersMutationRequestStatusEnumKey =
  (typeof postOrganizationsSlugSellersMutationRequestStatusEnum)[keyof typeof postOrganizationsSlugSellersMutationRequestStatusEnum];

export type PostOrganizationsSlugSellersMutationRequest = {
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  email: string;
  /**
   * @type string
   */
  phone: string;
  /**
   * @type string
   */
  companyName: string;
  /**
   * @type string
   */
  documentType: PostOrganizationsSlugSellersMutationRequestDocumentTypeEnumKey;
  /**
   * @type string
   */
  document: string;
  /**
   * @type string
   */
  country: string;
  /**
   * @type string
   */
  state: string;
  /**
   * @type string | undefined
   */
  city?: string;
  /**
   * @type string | undefined
   */
  street?: string;
  /**
   * @type string | undefined
   */
  zipCode?: string;
  /**
   * @type string | undefined
   */
  neighborhood?: string;
  /**
   * @type string | undefined
   */
  number?: string;
  /**
   * @type string | undefined
   */
  complement?: string;
  /**
   * @type string | undefined
   */
  status?: PostOrganizationsSlugSellersMutationRequestStatusEnumKey;
};

export type PostOrganizationsSlugSellersMutationResponse =
  PostOrganizationsSlugSellers201;

export type PostOrganizationsSlugSellersMutation = {
  Response: PostOrganizationsSlugSellers201;
  Request: PostOrganizationsSlugSellersMutationRequest;
  PathParams: PostOrganizationsSlugSellersPathParams;
  Errors: any;
};